+++ DocumentQuestionsFunction.sln
@@ -3,8 +3,8 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DocumentQuestionsFunction", "DocumentQuestionsFunction\DocumentQuestionsFunction.csproj", "{C4D5A230-367E-4C0A-A62A-C84997225B4E}"
-EndProject
+#Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DocumentQuestionsFunction", "DocumentQuestionsFunction\DocumentQuestionsFunction.csproj", "{C4D5A230-367E-4C0A-A62A-C84997225B4E}"
+#EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{7F25F21A-6BFF-4943-AF8E-A61242235AD6}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -35,10 +35,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Release|Any CPU.Build.0 = Release|Any CPU
+		#{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		#{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		#{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		#{C4D5A230-367E-4C0A-A62A-C84997225B4E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C201E846-3F44-41B6-865C-DD30F6462BDD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C201E846-3F44-41B6-865C-DD30F6462BDD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C201E846-3F44-41B6-865C-DD30F6462BDD}.Release|Any CPU.ActiveCfg = Release|Any CPU
