--- DocumentQuestionsFunction/DocumentQuestionsFunction.csproj
@@ -1,44 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-		<AzureFunctionsVersion>v4</AzureFunctionsVersion>
-		<OutputType>Exe</OutputType>
-		<PublishReadyToRun>true</PublishReadyToRun>
-		<RootNamespace>DocumentQuestions.Function</RootNamespace>
-	  <UserSecretsId>95c45569-a700-4ac8-9860-79dc624e4a44</UserSecretsId>
-  </PropertyGroup>
-	<ItemGroup>
-		<!--<PackageReference Include="Azure.Data.Tables" Version="12.9.1" />-->
-		<PackageReference Include="Azure.Identity" Version="1.17.0" />
-		<PackageReference Include="Azure.Monitor.OpenTelemetry.Exporter" Version="1.3.0" />
-      <PackageReference Include="Azure.Storage.Blobs" Version="12.26.0" />
-      <!--<PackageReference Include="Azure.Storage.Files.Shares" Version="12.21.0" />
-		<PackageReference Include="Azure.Storage.Queues" Version="12.21.0" />-->
-		<PackageReference Include="HttpMultipartParser" Version="9.0.0" />
-		<PackageReference Include="Microsoft.Azure.Functions.Extensions" Version="1.1.0" />
-		<PackageReference Include="Microsoft.Azure.Functions.Worker" Version="2.0.0" />
-		<PackageReference Include="Microsoft.Azure.Functions.Worker.Core" Version="2.0.0" />
-		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Abstractions" Version="1.3.0" />
-		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.3.0" />
-		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage" Version="6.7.0" />
-		<!--<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs" Version="6.6.0" />-->
-		<PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="2.0.2" />
-		<!--<PackageReference Include="Microsoft.Extensions.Azure" Version="1.9.0" />-->
-   </ItemGroup>
-	<ItemGroup>
-	  <ProjectReference Include="..\DocumentQuestionsLibrary\DocumentQuestionsLibrary.csproj" />
-	</ItemGroup>
-	<ItemGroup>
-		<None Update="appsettings.json">
-		  <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-		</None>
-		<None Update="host.json">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
-		<None Update="local.settings.json">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-			<CopyToPublishDirectory>Never</CopyToPublishDirectory>
-		</None>
-	</ItemGroup>
-
-</Project>
