--- DocumentQuestionsFunction/appsettings.json
@@ -1,24 +1,0 @@
-{
-   "IsEncrypted": false,
-   "Values": {
-      "AzureWebJobsStorage": "",
-      "OpenAIEndpoint": "",
-      "OpenAIKey": "",
-      "OpenAIChatModel": "gpt-4-32k",
-      "OpenAIChatDeploymentName": "gpt-4-32k",
-      "OpenAIEmbeddingModel": "text-embedding-ada-002",
-      "OpenAIEmbeddingDeploymentName": "text-embedding-ada-002",
-      "StorageAccount": "",
-      "StorageConnectionString": "",
-      "ExtractedContainerName": "extracted",
-      "ContainerName": "raw",
-      "RawStorageConnectionString": "",
-      "DocumentIntelligenceSubscriptionKey": "",
-      "DocumentIntelligenceEndpoint": "",
-      "CognitiveSearchEndpoint": "",
-      "CognitiveSearchAdminKey": "",
-      "UseOpenAIKey": "",
-      "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated"
-
-   }
-}
